Type Definition voicemeeter::bindings::T_VBVMR_VBAUDIOCALLBACK

pub type T_VBVMR_VBAUDIOCALLBACK = Option<unsafe extern "C" fn(lpUser: *mut c_void, nCommand: c_long, lpData: *mut c_void, nnn: c_long) -> c_long>;

@brief VB-AUDIO Callback is called for different task to Initialize, perform and end your process. VB-AUDIO Callback is part of single TIME CRITICAL Thread. VB-AUDIO Callback is non re-entrant (cannot be called while in process) VB-AUDIO Callback is supposed to be REAL TIME when called to process buffer. (it means that the process has to be performed as fast as possible, waiting cycles are forbidden. do not use O/S synchronization object, even Critical_Section can generate waiting cycle. Do not use system functions that can generate waiting cycle like display, disk or communication functions for example).

@param lpUser: User pointer given on callback registration. @param ncommand: reason why the callback is called. @param lpData: pointer on structure, pending on nCommand. @param nnn: additional data, unused

@return : 0: always 0 (unused).