voicemeeter

Enum CallbackCommand

#[repr(i32)]
pub enum CallbackCommand<'a> {
    Starting(Starting<'a>),
    Ending(Ending<'a>),
    Change(Change<'a>),
    BufferIn(BufferIn<'a>),
    BufferOut(BufferOut<'a>),
    BufferMain(BufferMain<'a>),
    Other(VBVMR_CBCOMMAND, RawCallbackData),
}

Available on crate feature interface only.

Callback command passed to the audio callback.

Variants

Starting(Starting<'a>)

Starting command

this is the first call of your Callback, made to let you initialize your possible different DSP processing objects, allocate memory, precompute date…

Ending(Ending<'a>)

Ending command

this is the last call of your Callback, to release all your structure previously allocated in the first call.

Change(Change<'a>)

Change command

this command is called if the samplerate or buffer size have changed

BufferIn(BufferIn<'a>)

BufferIn command

buffers for inputs

BufferOut(BufferOut<'a>)

BufferOut command

buffers for outputs

BufferMain(BufferMain<'a>)

BufferMain command

buffers for all i/o

Other(VBVMR_CBCOMMAND, RawCallbackData)

Other inknown command

Implementations

impl<'a> CallbackCommand<'a>

pub fn name(&self) -> Option<&'static str>

Get the command “name”