voicemeeter::interface

Module get_levels

Available on crate feature interface only.

Level and midi related functions

Functions

get_level

get_midi_message

get_midi_message_buff

Re-exports

pub use crate::types::Device;

pub use crate::types::LevelType;

Enums

GetLevelError

Errors that can happen when querying levels from Voicemeeter.

GetMidiMessageError

Errors that can happen when querying midi messages from Voicemeeter.