Module voicemeeter::bindings

Raw FFI Bindings

Structs

VBVMR_AUDIOCALLBACK

VBVMR_CBCOMMAND

VBVMR_DEVTYPE

@name Device Enumeration Functions @{

VBVMR_MACROBUTTON_MODE

VMRTSTATE_MODE

VoicemeeterRemoteRaw

tagVBAN_VMRT_PACKET

tagVBVMR_AUDIOBUFFER

tagVBVMR_AUDIOINFO

@name VB-Audio Callback Functions @{

tagVBVMR_INTERFACE

Constants

VBVMR_RESULT_OK

expected_size_T_VBAN_VMRT_PACKET

Type Aliases

LPT_VBAN_VMRT_PACKET

LPT_VBVMR_INTERFACE

PT_VBAN_VMRT_PACKET

PT_VBVMR_INTERFACE

T_VBAN_VMRT_PACKET

T_VBVMR_AudioCallbackRegister

T_VBVMR_AudioCallbackStart

T_VBVMR_AudioCallbackStop

T_VBVMR_AudioCallbackUnregister

T_VBVMR_GetLevel

T_VBVMR_GetMidiMessage

T_VBVMR_GetParameterFloat

T_VBVMR_GetParameterStringA

T_VBVMR_GetParameterStringW

T_VBVMR_GetVoicemeeterType

T_VBVMR_GetVoicemeeterVersion

T_VBVMR_INTERFACE

T_VBVMR_Input_GetDeviceDescA

T_VBVMR_Input_GetDeviceDescW

T_VBVMR_Input_GetDeviceNumber

T_VBVMR_IsParametersDirty

T_VBVMR_Login

T_VBVMR_Logout

T_VBVMR_MacroButton_GetStatus

T_VBVMR_MacroButton_IsDirty

T_VBVMR_MacroButton_SetStatus

T_VBVMR_Output_GetDeviceDescA

T_VBVMR_Output_GetDeviceDescW

T_VBVMR_Output_GetDeviceNumber

T_VBVMR_RunVoicemeeter

T_VBVMR_SetParameterFloat

T_VBVMR_SetParameterStringA

T_VBVMR_SetParameterStringW

T_VBVMR_SetParameters

T_VBVMR_SetParametersW

T_VBVMR_VBAUDIOCALLBACK

@brief VB-AUDIO Callback is called for different task to Initialize, perform and end your process. VB-AUDIO Callback is part of single TIME CRITICAL Thread. VB-AUDIO Callback is non re-entrant (cannot be called while in process) VB-AUDIO Callback is supposed to be REAL TIME when called to process buffer. (it means that the process has to be performed as fast as possible, waiting cycles are forbidden. do not use O/S synchronization object, even Critical_Section can generate waiting cycle. Do not use system functions that can generate waiting cycle like display, disk or communication functions for example).

VBVMR_LPT_AUDIOBUFFER

VBVMR_LPT_AUDIOINFO

@name VB-Audio Callback Functions @{

VBVMR_PT_AUDIOBUFFER

VBVMR_PT_AUDIOINFO

@name VB-Audio Callback Functions @{

VBVMR_T_AUDIOBUFFER

VBVMR_T_AUDIOINFO

@name VB-Audio Callback Functions @{