Struct voicemeeter::bindings::VoicemeeterRemoteRaw

pub struct VoicemeeterRemoteRaw {

    pub VBVMR_Login: unsafe extern "C" fn() -> c_long,
    pub VBVMR_Logout: unsafe extern "C" fn() -> c_long,
    pub VBVMR_RunVoicemeeter: unsafe extern "C" fn(vType: c_long) -> c_long,
    pub VBVMR_GetVoicemeeterType: unsafe extern "C" fn(pType: *mut c_long) -> c_long,
    pub VBVMR_GetVoicemeeterVersion: unsafe extern "C" fn(pVersion: *mut c_long) -> c_long,
    pub VBVMR_IsParametersDirty: unsafe extern "C" fn() -> c_long,
    pub VBVMR_GetParameterFloat: unsafe extern "C" fn(szParamName: *mut c_char, pValue: *mut f32) -> c_long,
    pub VBVMR_GetParameterStringA: unsafe extern "C" fn(szParamName: *mut c_char, szString: *mut c_char) -> c_long,
    pub VBVMR_GetParameterStringW: unsafe extern "C" fn(szParamName: *mut c_char, wszString: *mut c_ushort) -> c_long,
    pub VBVMR_GetLevel: unsafe extern "C" fn(nType: c_long, nuChannel: c_long, pValue: *mut f32) -> c_long,
    pub VBVMR_GetMidiMessage: unsafe extern "C" fn(pMIDIBuffer: *mut c_uchar, nbByteMax: c_long) -> c_long,
    pub VBVMR_SetParameterFloat: unsafe extern "C" fn(szParamName: *mut c_char, Value: f32) -> c_long,
    pub VBVMR_SetParameterStringA: unsafe extern "C" fn(szParamName: *mut c_char, szString: *mut c_char) -> c_long,
    pub VBVMR_SetParameterStringW: unsafe extern "C" fn(szParamName: *mut c_char, wszString: *mut c_ushort) -> c_long,
    pub VBVMR_SetParameters: unsafe extern "C" fn(szParamScript: *mut c_char) -> c_long,
    pub VBVMR_SetParametersW: unsafe extern "C" fn(szParamScript: *mut c_ushort) -> c_long,
    pub VBVMR_Output_GetDeviceNumber: unsafe extern "C" fn() -> c_long,
    pub VBVMR_Output_GetDeviceDescA: unsafe extern "C" fn(zindex: c_long, nType: *mut c_long, szDeviceName: *mut c_char, szHardwareId: *mut c_char) -> c_long,
    pub VBVMR_Output_GetDeviceDescW: unsafe extern "C" fn(zindex: c_long, nType: *mut c_long, wszDeviceName: *mut c_ushort, wszHardwareId: *mut c_ushort) -> c_long,
    pub VBVMR_Input_GetDeviceNumber: unsafe extern "C" fn() -> c_long,
    pub VBVMR_Input_GetDeviceDescA: unsafe extern "C" fn(zindex: c_long, nType: *mut c_long, szDeviceName: *mut c_char, szHardwareId: *mut c_char) -> c_long,
    pub VBVMR_Input_GetDeviceDescW: unsafe extern "C" fn(zindex: c_long, nType: *mut c_long, wszDeviceName: *mut c_ushort, wszHardwareId: *mut c_ushort) -> c_long,
    pub VBVMR_AudioCallbackRegister: unsafe extern "C" fn(mode: c_long, pCallback: T_VBVMR_VBAUDIOCALLBACK, lpUser: *mut c_void, szClientName: *mut c_char) -> c_long,
    pub VBVMR_AudioCallbackStart: unsafe extern "C" fn() -> c_long,
    pub VBVMR_AudioCallbackStop: unsafe extern "C" fn() -> c_long,
    pub VBVMR_AudioCallbackUnregister: unsafe extern "C" fn() -> c_long,
    pub VBVMR_MacroButton_IsDirty: unsafe extern "C" fn() -> c_long,
    pub VBVMR_MacroButton_GetStatus: unsafe extern "C" fn(nuLogicalButton: c_long, pValue: *mut f32, bitmode: c_long) -> c_long,
    pub VBVMR_MacroButton_SetStatus: unsafe extern "C" fn(nuLogicalButton: c_long, fValue: f32, bitmode: c_long) -> c_long,
    /* private fields */

}

Fields
VBVMR_Login: unsafe extern "C" fn() -> c_longVBVMR_Logout: unsafe extern "C" fn() -> c_longVBVMR_RunVoicemeeter: unsafe extern "C" fn(vType: c_long) -> c_longVBVMR_GetVoicemeeterType: unsafe extern "C" fn(pType: *mut c_long) -> c_longVBVMR_GetVoicemeeterVersion: unsafe extern "C" fn(pVersion: *mut c_long) -> c_longVBVMR_IsParametersDirty: unsafe extern "C" fn() -> c_longVBVMR_GetParameterFloat: unsafe extern "C" fn(szParamName: *mut c_char, pValue: *mut f32) -> c_longVBVMR_GetParameterStringA: unsafe extern "C" fn(szParamName: *mut c_char, szString: *mut c_char) -> c_longVBVMR_GetParameterStringW: unsafe extern "C" fn(szParamName: *mut c_char, wszString: *mut c_ushort) -> c_longVBVMR_GetLevel: unsafe extern "C" fn(nType: c_long, nuChannel: c_long, pValue: *mut f32) -> c_longVBVMR_GetMidiMessage: unsafe extern "C" fn(pMIDIBuffer: *mut c_uchar, nbByteMax: c_long) -> c_longVBVMR_SetParameterFloat: unsafe extern "C" fn(szParamName: *mut c_char, Value: f32) -> c_longVBVMR_SetParameterStringA: unsafe extern "C" fn(szParamName: *mut c_char, szString: *mut c_char) -> c_longVBVMR_SetParameterStringW: unsafe extern "C" fn(szParamName: *mut c_char, wszString: *mut c_ushort) -> c_longVBVMR_SetParameters: unsafe extern "C" fn(szParamScript: *mut c_char) -> c_longVBVMR_SetParametersW: unsafe extern "C" fn(szParamScript: *mut c_ushort) -> c_longVBVMR_Output_GetDeviceNumber: unsafe extern "C" fn() -> c_longVBVMR_Output_GetDeviceDescA: unsafe extern "C" fn(zindex: c_long, nType: *mut c_long, szDeviceName: *mut c_char, szHardwareId: *mut c_char) -> c_longVBVMR_Output_GetDeviceDescW: unsafe extern "C" fn(zindex: c_long, nType: *mut c_long, wszDeviceName: *mut c_ushort, wszHardwareId: *mut c_ushort) -> c_longVBVMR_Input_GetDeviceNumber: unsafe extern "C" fn() -> c_longVBVMR_Input_GetDeviceDescA: unsafe extern "C" fn(zindex: c_long, nType: *mut c_long, szDeviceName: *mut c_char, szHardwareId: *mut c_char) -> c_longVBVMR_Input_GetDeviceDescW: unsafe extern "C" fn(zindex: c_long, nType: *mut c_long, wszDeviceName: *mut c_ushort, wszHardwareId: *mut c_ushort) -> c_longVBVMR_AudioCallbackRegister: unsafe extern "C" fn(mode: c_long, pCallback: T_VBVMR_VBAUDIOCALLBACK, lpUser: *mut c_void, szClientName: *mut c_char) -> c_longVBVMR_AudioCallbackStart: unsafe extern "C" fn() -> c_longVBVMR_AudioCallbackStop: unsafe extern "C" fn() -> c_longVBVMR_AudioCallbackUnregister: unsafe extern "C" fn() -> c_longVBVMR_MacroButton_IsDirty: unsafe extern "C" fn() -> c_longVBVMR_MacroButton_GetStatus: unsafe extern "C" fn(nuLogicalButton: c_long, pValue: *mut f32, bitmode: c_long) -> c_longVBVMR_MacroButton_SetStatus: unsafe extern "C" fn(nuLogicalButton: c_long, fValue: f32, bitmode: c_long) -> c_long
Implementations

impl VoicemeeterRemoteRaw

pub unsafe fn new<P>(path: P) -> Result<Self, Error> where
P: AsRef<OsStr>,

pub unsafe fn from_library<L>(library: L) -> Result<Self, Error> where
L: Into<Library>,

pub unsafe fn VBVMR_Login(&self) -> c_long

@brief Open Communication Pipe With Voicemeeter (typically called on software startup). @return : 0: OK (no error). 1: OK but Voicemeeter Application not launched. -1: cannot get client (unexpected) -2: unexpected login (logout was expected before).

pub unsafe fn VBVMR_Logout(&self) -> c_long

@brief Close Communication Pipe With Voicemeeter (typically called on software end). @return : 0 if ok.

pub unsafe fn VBVMR_RunVoicemeeter(&self, vType: c_long) -> c_long

@brief Run Voicemeeter Application (get installation directory and run Voicemeeter Application). @param vType : Voicemeeter type (1 = Voicemeeter, 2= Voicemeeter Banana, 3= Voicemeeter Potato, 6 = Potato x64 bits). @return : 0: Ok. -1: not installed (UninstallString not found in registry). -2: unknown vType number

pub unsafe fn VBVMR_GetVoicemeeterType(&self, pType: *mut c_long) -> c_long

@brief Get Voicemeeter Type @param pType : Pointer on 32bit long receiving the type (1 = Voicemeeter, 2= Voicemeeter Banana, 3 Potato).

VOICEMEETER STRIP/BUS INDEX ASSIGNMENT

| Strip 1 | Strip 2 |Virtual Input| BUS A | BUS B | +———+———+———––+———+———+ | 0 | 1 | 2 | 0 | 1 |

VOICEMEETER BANANA STRIP/BUS INDEX ASSIGNMENT

| Strip 1 | Strip 2 | Strip 2 |Virtual Input|Virtual AUX|BUS A1|BUS A2|BUS A3|BUS B1|BUS B2| +———+———+———+———––+———–+——+——+——+——+——+ | 0 | 1 | 2 | 3 | 4 | 0 | 1 | 2 | 3 | 4 |

VOICEMEETER POTATO STRIP/BUS INDEX ASSIGNMENT

| Strip 1 | Strip 2 | Strip 2 | Strip 2 | Strip 2 |Virtual Input|Virtual AUX| VAIO3 |BUS A1|BUS A2|BUS A3|BUS A4|BUS A5|BUS B1|BUS B2|BUS B3| +———+———+———+———+———+———––+———–+———–+——+——+——+——+——+——+——+——+ | 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 |

@return : 0: OK (no error). -1: cannot get client (unexpected) -2: no server.

pub unsafe fn VBVMR_GetVoicemeeterVersion(
&self,
pVersion: *mut c_long
) -> c_long

@brief Get Voicemeeter Version @param pType : Pointer on 32bit integer receiving the version (v1.v2.v3.v4) v1 = (version & 0xFF000000)>>24; v2 = (version & 0x00FF0000)>>16; v3 = (version & 0x0000FF00)>>8; v4 = version & 0x000000FF;

@return : 0: OK (no error). -1: cannot get client (unexpected) -2: no server.

pub unsafe fn VBVMR_IsParametersDirty(&self) -> c_long

@brief Check if parameters have changed. Call this function periodically (typically every 10 or 20ms). (this function must be called from one thread only)

@return: 0: no new paramters. 1: New parameters -> update your display. -1: error (unexpected) -2: no server.

pub unsafe fn VBVMR_GetParameterFloat(
&self,
szParamName: *mut c_char,
pValue: *mut f32
) -> c_long

@brief get parameter value. @param szParamName : Null Terminal ASCII String giving the name of the parameter (see parameters name table) @param pValue : Pointer on float (32bit float by reference) receiving the wanted value. @return : 0: OK (no error). -1: error -2: no server. -3: unknown parameter -5: structure mismatch

pub unsafe fn VBVMR_GetParameterStringA(
&self,
szParamName: *mut c_char,
szString: *mut c_char
) -> c_long

@brief get parameter value. @param szParamName : Null Terminal ASCII String giving the name of the parameter (see parameters name table) @param pValue : Pointer on String (512 char or wchar) receiving the wanted value. @return : 0: OK (no error). -1: error -2: no server. -3: unknown parameter -5: structure mismatch

pub unsafe fn VBVMR_GetParameterStringW(
&self,
szParamName: *mut c_char,
wszString: *mut c_ushort
) -> c_long

pub unsafe fn VBVMR_GetLevel(
&self,
nType: c_long,
nuChannel: c_long,
pValue: *mut f32
) -> c_long

@brief Get Current levels. (this function must be called from one thread only)

@param nType: 0= pre fader input levels. 1= post fader input levels. 2= post Mute input levels. 3= output levels.

@param nuChannel: audio channel zero based index for input 0 = in#1 left, 1= in#1 Right, etc… for output 0 = busA ch1, 1 = busA ch2…

VOICEMEETER CHANNEL ASSIGNMENT

| Strip 1 | Strip 2 | Virtual Input | +––+––+––+––+––+––+––+––+––+––+––+––+ | 00 | 01 | 02 | 03 | 04 | 05 | 06 | 07 | 08 | 09 | 10 | 11 |

| Output A1 / A2 | Virtual Output | +––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+ | 00 | 01 | 02 | 03 | 04 | 05 | 06 | 07 | 08 | 09 | 10 | 11 | 12 | 13 | 14 | 15 |

VOICEMEETER BANANA CHANNEL ASSIGNMENT

| Strip 1 | Strip 2 | Strip 3 | Virtual Input | Virtual Input AUX | +––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+ | 00 | 01 | 02 | 03 | 04 | 05 | 06 | 07 | 08 | 09 | 10 | 11 | 12 | 13 | 14 | 15 | 16 | 17 | 18 | 19 | 20 | 21 |

| Output A1 | Output A2 | Output A3 | +––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+ | 00 | 01 | 02 | 03 | 04 | 05 | 06 | 07 | 08 | 09 | 10 | 11 | 12 | 13 | 14 | 15 | 16 | 17 | 18 | 19 | 20 | 21 | 22 | 23 |

| Virtual Output B1 | Virtual Output B2 | +––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+ | 24 | 25 | 26 | 27 | 28 | 29 | 30 | 31 | 32 | 33 | 34 | 35 | 36 | 37 | 38 | 39 |

VOICEMEETER POTATO CHANNEL ASSIGNMENT

| Strip 1 | Strip 2 | Strip 3 | Strip 4 | Strip 5 | Virtual Input | Virtual Input AUX | VAIO3 | +––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+ | 00 | 01 | 02 | 03 | 04 | 05 | 06 | 07 | 08 | 09 | 10 | 11 | 12 | 13 | 14 | 15 | 16 | 17 | 18 | 19 | 20 | 21 | 22 | 23 | 25 | 25 | 26 | 27 | 28 | 29 | 30 | 31 | 32 | 33 |

| Output A1 | Output A2 | Output A3 | Output A4 | Output A5 | +––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+ | 00 | 01 | 02 | 03 | 04 | 05 | 06 | 07 | 08 | 09 | 10 | 11 | 12 | 13 | 14 | 15 | 16 | 17 | 18 | 19 | 20 | 21 | 22 | 23 | 24 | 25 | 26 | 27 | 28 | 29 | 30 | 31 | 32 | 33 | 34 | 35 | 36 | 37 | 38 | 39 |

| Virtual Output B1 | Virtual Output B2 | Virtual Output B3 | +––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+––+ | 40 | 41 | 42 | 43 | 44 | 45 | 46 | 47 | 48 | 49 | 50 | 51 | 52 | 53 | 54 | 55 | 56 | 57 | 58 | 59 | 60 | 61 | 62 | 63 |

@return : 0: OK (no error). -1: error -2: no server. -3: no level available -4: out of range

pub unsafe fn VBVMR_GetMidiMessage(
&self,
pMIDIBuffer: *mut c_uchar,
nbByteMax: c_long
) -> c_long

@brief Get MIDI message from M.I.D.I. input device used by Voicemeeter M.I.D.I. mapping. (this function must be called from one thread only)

@param pMIDIBuffer: pointer on MIDI Buffer. Expected message size is below 4 bytes, but it’s recommended to use 1024 Bytes local buffer to receive possible multiple M.I.D.I. event message in optimal way: unsigned char pBuffer[1024];

@return : >0: number of bytes placed in buffer (2 or 3 byte for usual M.I.D.I. message) -1: error -2: no server. -5: no MIDI data -6: no MIDI data

pub unsafe fn VBVMR_SetParameterFloat(
&self,
szParamName: *mut c_char,
Value: f32
) -> c_long

@brief Set a single float 32 bits parameters . @param szParamName : Null Terminal ASCII String giving the name of the parameter (see parameters name table) example: Strip[1].gain Strip[0].mute Bus[0].gain Bus[0].eq.channel[0].cell[0].gain

@param pValue : float 32bit containing the new value. @return : 0: OK (no error). -1: error -2: no server. -3: unknown parameter

pub unsafe fn VBVMR_SetParameterStringA(
&self,
szParamName: *mut c_char,
szString: *mut c_char
) -> c_long

@brief Set a single string parameters . @param szParamName : Null Terminal ASCII String giving the name of the parameter (see parameters name table) example: Strip[1].name Strip[0].device.mme Bus[0].device.asio

@param szString : zero terminal string. @return : 0: OK (no error). -1: error -2: no server. -3: unknown parameter

pub unsafe fn VBVMR_SetParameterStringW(
&self,
szParamName: *mut c_char,
wszString: *mut c_ushort
) -> c_long

pub unsafe fn VBVMR_SetParameters(&self, szParamScript: *mut c_char) -> c_long

@brief Set one or several parameters by a script ( < 48 kB ). @param szParamName : Null Terminal ASCII String giving the script (script allows to change several parameters in the same time - SYNCHRO). Possible Instuction separators: ‘,’ ‘;’ or ‘\n’(CR) EXAMPLE: “Strip[0].gain = -6.0 Strip[0].A1 = 0 Strip[0].B1 = 1 Strip[1].gain = -6.0 Strip[2].gain = 0.0 Strip[3].name = “Skype Caller” “

@return : 0: OK (no error). >0: number of line causing script error. -1: error -2: no server. -3: unexpected error -4: unexpected error

pub unsafe fn VBVMR_SetParametersW(
&self,
szParamScript: *mut c_ushort
) -> c_long

pub unsafe fn VBVMR_Output_GetDeviceNumber(&self) -> c_long

@brief Get number of Audio Output Device available on the system @return : return number of device found.

pub unsafe fn VBVMR_Output_GetDeviceDescA(
&self,
zindex: c_long,
nType: *mut c_long,
szDeviceName: *mut c_char,
szHardwareId: *mut c_char
) -> c_long

@brief Return pointer on Output Device Descriptor according index @param zindex : zero based index @param nType : Pointer on 32bit long receiving the type (pointer can be NULL). @param szName : Pointer on string (256 char min) receiving the device name (pointer can be NULL). @param szHardwareId : Pointer on string (256 char min) receiving the hardware ID (pointer can be NULL). @return : 0: OK (no error).

pub unsafe fn VBVMR_Output_GetDeviceDescW(
&self,
zindex: c_long,
nType: *mut c_long,
wszDeviceName: *mut c_ushort,
wszHardwareId: *mut c_ushort
) -> c_long

pub unsafe fn VBVMR_Input_GetDeviceNumber(&self) -> c_long

@brief Get number of Audio Input Device available on the system @return : return number of device found.

pub unsafe fn VBVMR_Input_GetDeviceDescA(
&self,
zindex: c_long,
nType: *mut c_long,
szDeviceName: *mut c_char,
szHardwareId: *mut c_char
) -> c_long

@brief Return pointer on Input Device Descriptor according index @param zindex : zero based index @param nType : Pointer on 32bit long receiving the type (pointer can be NULL). @param szName : Pointer on string (256 char min) receiving the device name (pointer can be NULL). @param szHardwareId : Pointer on string (256 char min) receiving the hardware ID (pointer can be NULL). @return : 0: OK (no error).

pub unsafe fn VBVMR_Input_GetDeviceDescW(
&self,
zindex: c_long,
nType: *mut c_long,
wszDeviceName: *mut c_ushort,
wszHardwareId: *mut c_ushort
) -> c_long

pub unsafe fn VBVMR_AudioCallbackRegister(
&self,
mode: c_long,
pCallback: T_VBVMR_VBAUDIOCALLBACK,
lpUser: *mut c_void,
szClientName: *mut c_char
) -> c_long

@brief register your audio callback function to receive real time audio buffer it’s possible to register up to 3x different Audio Callback in the same application or in 3x different applications. In the same application, this is possible because Voicemeeter provides 3 kind of audio Streams:

AUDIO INPUT INSERT (to process all Voicemeeter inputs as insert)

AUDIO OUTPUT INSERT (to process all Voicemeeter BUS outputs as insert)

ALL AUDIO I/O (to process all Voicemeeter i/o). Note: a single callback can be used to receive the 3 possible audio streams.

@param mode : callback type (main, input or bus output) see define below @param pCallback : Pointer on your callback function. @param lpUser : user pointer (pointer that will be passed in callback first argument). @param szClientName[64]: IN: Name of the application registering the Callback. OUT: Name of the application already registered. @return : 0: OK (no error). -1: error 1: callback already registered (by another application).

pub unsafe fn VBVMR_AudioCallbackStart(&self) -> c_long

@brief Start / Stop Audio processing

he Callback will be called with @return : 0: OK (no error). -1: error -2: no callback registred.

pub unsafe fn VBVMR_AudioCallbackStop(&self) -> c_long

pub unsafe fn VBVMR_AudioCallbackUnregister(&self) -> c_long

@brief unregister your callback to release voicemeeter virtual driver (this function will automatically call VBVMR_AudioCallbackStop() function) @param pCallback : Pointer on your callback function. @return : 0: OK (no error). -1: error 1: callback already unregistered.

pub unsafe fn VBVMR_MacroButton_IsDirty(&self) -> c_long

@brief Check if Macro Buttons states changed. Call this function periodically (typically every 50 or 500ms) to know if something happen on MacroButton states . (this function must be called from one thread only)

@return: 0: no new status. >0: last nu logical button status changed. -1: error (unexpected) -2: no server.

pub unsafe fn VBVMR_MacroButton_GetStatus(
&self,
nuLogicalButton: c_long,
pValue: *mut f32,
bitmode: c_long
) -> c_long

@brief get current status of a given button. @param nuLogicalButton : button index: 0 to 79) @param pValue : Pointer on float (32bit float by reference) receiving the wanted value (0.0 = OFF / 1.0 = ON). @param bitmode: define what kind of value you want to read (see MACROBUTTON_MODE below) @return : 0: OK (no error). -1: error -2: no server. -3: unknown parameter -5: structure mismatch

pub unsafe fn VBVMR_MacroButton_SetStatus(
&self,
nuLogicalButton: c_long,
fValue: f32,
bitmode: c_long
) -> c_long

@brief set current button value. @param nuLogicalButton : button index: 0 to 79) @param fValue : float 32 bit value giving the status (0.0 = OFF / 1.0 = ON). @param bitmode: define what kind of value you want to write/modify (see MACROBUTTON_MODE below) @return : 0: OK (no error). -1: error -2: no server. -3: unknown parameter -5: structure mismatch