Module voicemeeter::interface::parameters

Available on crate feature interface only.

Functions for setting and getting parameter values.

Note that if your application exits quickly after setting a parameter, voicemeeter may not update. If you do this, you should maybe insert a small sleep before dropping the remote.’

Functions

parameters

Re-exports

pub use bus::*;

pub use eq::*;

pub use fx::*;

pub use option::*;

pub use recorder::*;

pub use strip::*;

pub use vban::*;

Modules

bus

Bus parameters

eq

Common structs for EQ

fx

Fx

get_parameters

Parameter related functions and types

option

Option parameters

recorder

Recorder

set_parameters

Functions and data types for setting parameter values.

strip

Strip parameters

vban

VBAN

Structs

BoolParameter

A boolean parameter

DeviceError

Device is invalid for the current program and parameter

FloatParameter

A float parameter

IntParameter

A integer parameter

InvalidVoicemeeterVersion

Invalid Voicemeeter program

OutOfRangeError

Parameter is out of range for current program

Parameters

Parameter abstraction

StringParameter

A string parameter

TupleParameter

A tuple parameter

Enums

InvalidTypeError

Invalid strip/bus type for a specific parameter

ParameterError

Invalid Parameter

Traits

BusIndex

Bus index helper

StripIndex

Strip index helper