Struct voicemeeter::interface::parameters::strip::Strip

pub struct Strip<'a> { /* private fields */ }

Available on crate feature interface only.

Parameters for a strip.

A strip is a physical or virtual input.

Returned by VoicemeeterRemote::parameters().strip(i)

Example

use voicemeeter::VoicemeeterRemote;

// Get the client.
let remote = VoicemeeterRemote::new()?;

// Get the label of strip 1 (index 0)
println!("{}", remote.parameters().strip(0)?.label().get()?);
// Set strip 3 (index 2) to output to A1
remote.parameters().strip(2)?.a1().set(true)?;

// Ensure the change is registered.
remote.is_parameters_dirty()?;

// We need to sleep here because otherwise the change won't be registered,
// in a long running program this is not needed.
std::thread::sleep(std::time::Duration::from_millis(50));

Implementations

impl<'a> Strip<'a>

pub fn param(&self, dot: impl Display) -> Cow<'static, ParameterNameRef>

Get the identifier for a parameter on this strip: Strip[i].{dot}

pub fn is_physical(&self) -> bool

Strip is physical

pub fn is_virtual(&self) -> bool

Strip is virtual

pub fn mono(&self) -> BoolParameter<'_>

Mono Button

pub fn mute(&self) -> BoolParameter<'_>

Mute Button

pub fn solo(&self) -> BoolParameter<'_>

Solo Button

pub fn mute_center(&self) -> BoolParameter<'_>

Mute Center Button

pub fn gain(&self) -> FloatParameter<'_>

Gain slider

pub fn gain_layer(&self, layer: impl Into<ZIndex>) -> FloatParameter<'_>

Gain slider for a bus

pub fn pan_x(&self) -> FloatParameter<'_>

Pan in x direction

pub fn pan_y(&self) -> FloatParameter<'_>

Pan in y direction

pub fn color_x(&self) -> Result<FloatParameter<'_>, InvalidTypeError>

Color of physical strip in x direction

pub fn color_y(&self) -> Result<FloatParameter<'_>, InvalidTypeError>

Color of physical strip in y direction

pub fn fx_x(&self) -> Result<FloatParameter<'_>, InvalidTypeError>

FX of physical strip in x direction

pub fn fx_y(&self) -> Result<FloatParameter<'_>, InvalidTypeError>

FX of physical strip in y direction

pub fn audability(&self) -> FloatParameter<'_>

Audability

pub fn comp(&self) -> FloatParameter<'_>

Compression

See also Strip::comp_detailed for detailed compressor settings

pub fn comp_detailed(&self) -> Result<StripCompressor<'_>, ParameterError>

Compressor detailed parameters/settings

Only works on Voicemeeter Potato

pub fn gate(&self) -> FloatParameter<'_>

Gate

See also Strip::gate_detailed for detailed gate settings

pub fn gate_detailed(&self) -> Result<StripGate<'_>, ParameterError>

Gate detailed parameters/settings

Only works on Voicemeeter Potato

pub fn denoiser(&self) -> Result<FloatParameter<'_>, ParameterError>

Denoiser Knob

pub fn karaoke(&self) -> IntParameter<'_>

Karaoke

pub fn limit(&self) -> IntParameter<'_>

Limit

pub fn eq_gain1(&self) -> Result<FloatParameter<'_>, InvalidTypeError>

EQGain1 of virtual strip

pub fn eq_gain2(&self) -> Result<FloatParameter<'_>, InvalidTypeError>

EQGain2 of virtual strip

pub fn eq_gain3(&self) -> Result<FloatParameter<'_>, InvalidTypeError>

EQGain3 of virtual strip

pub fn label(&self) -> StringParameter<'_>

Label

pub fn a1(&self) -> BoolParameter<'_>

Out BUS Assignation for A1

pub fn a2(&self) -> BoolParameter<'_>

Out BUS Assignation for A2

pub fn a3(&self) -> BoolParameter<'_>

Out BUS Assignation for A3

pub fn a4(&self) -> BoolParameter<'_>

Out BUS Assignation for A4

pub fn a5(&self) -> BoolParameter<'_>

Out BUS Assignation for A5

pub fn b1(&self) -> BoolParameter<'_>

Out BUS Assignation for B1

pub fn b2(&self) -> BoolParameter<'_>

Out BUS Assignation for B2

pub fn b3(&self) -> BoolParameter<'_>

Out BUS Assignation for B3

pub fn eq_on(&self) -> BoolParameter<'_>

EQ Button

pub fn eq_ab(&self) -> BoolParameter<'_>

EQ Memory Slot

pub fn eq( &self, channel: usize ) -> Result<EqChannelParameter<'_>, ParameterError>

EQ on channel

pub fn fade_to(&self) -> TupleParameter<'_, i32, usize>

Fade to

pub fn fade_by(&self) -> TupleParameter<'_, i32, usize>

Fade by

pub fn reverb(&self) -> FloatParameter<'_>

Send Level To Reverb

pub fn delay(&self) -> FloatParameter<'_>

Send Level To Delay

pub fn fx1(&self) -> FloatParameter<'_>

Send Level To External Fx1

pub fn fx2(&self) -> FloatParameter<'_>

Send Level To External Fx2

pub fn post_reverb(&self) -> BoolParameter<'_>

Post Reverb button

pub fn post_delay(&self) -> BoolParameter<'_>

Post Delay button

pub fn post_fx1(&self) -> BoolParameter<'_>

Post Fx1 button

pub fn post_fx2(&self) -> BoolParameter<'_>

Post Fx2 button

pub fn app_gain_indexed( &self, application_index: ZIndex ) -> FloatParameter<'_, true, false>

Application gain

pub fn app_mute_indexed( &self, application_index: ZIndex ) -> BoolParameter<'_, true, false>

Application Mute

pub fn app_gain(&self) -> TupleParameter<'_, String, f32, true, false>

Application gain

pub fn app_mute(&self) -> TupleParameter<'_, String, bool, true, false>

Application Mute

pub fn device(&self) -> Result<StripDevice<'a>, InvalidTypeError>

Audio Device information