Enum voicemeeter::types::Device

pub enum Device {

    Strip1,
    Strip2,
    Strip3,
    Strip4,
    Strip5,
    OutputA1,
    OutputA2,
    OutputA3,
    OutputA4,
    OutputA5,
    VirtualOutputB1,
    VirtualOutputB2,
    VirtualOutputB3,
    VirtualInput,
    VirtualInputAux,
    VirtualInput8,

}

Available on crate feature interface only.

A device.

Variants

Strip1

Input Strip 1. Available on all Voicemeeter versions.

Strip2

Input Strip 2. Available on all Voicemeeter versions.

Strip3

Input Strip 3. Available on Voicemeeter Banana and Potato.

Strip4

Input Strip 4. Available on Voicemeeter Potato.

Strip5

Input Strip 5. Available on Voicemeeter Potato.

OutputA1

Output A1. Available on all Voicemeeter versions.

OutputA2

Output A2. Available on all Voicemeeter versions.

OutputA3

Output A3. Available on Voicemeeter Banana and Potato.

OutputA4

Output A4. Available on Voicemeeter Potato.

OutputA5

Output A5. Available on Voicemeeter Potato.

VirtualOutputB1

Virtual Output B1. Available on all Voicemeeter versions.

VirtualOutputB2

Virtual Output B2. Available on Voicemeeter Banana and Potato.

VirtualOutputB3

Virtual Output B3. Available on Voicemeeter Potato.

VirtualInput

Virtual Input. Available on all Voicemeeter versions.

VirtualInputAux

Virtual Input Aux. Available on Voicemeeter Banana and Potato.

VirtualInput8

Virtual Input8. Available on Voicemeeter Potato.

Implementations

impl Device

pub const fn main( &self, program: &VoicemeeterApplication ) -> (Option<ChannelIndex>, Option<ChannelIndex>)

Get the ChannelIndex for this channel in the buffers when in main mode, if available in the current program.

pub const fn input( &self, program: &VoicemeeterApplication ) -> Option<ChannelIndex>

Get the ChannelIndex for this channel in the buffers when in input mode, if available in the current program.

pub const fn output( &self, program: &VoicemeeterApplication ) -> Option<ChannelIndex>

Get the ChannelIndex for this channel in the buffers when in output mode, if available in the current program.

pub const fn all() -> &'static [Self]

Get all channels available.

pub const fn as_strip_index( &self, program: &VoicemeeterApplication ) -> Option<ZIndex>

Get the strip index for this device in the current program.

pub const fn as_bus_index( &self, program: &VoicemeeterApplication ) -> Option<(ZIndex, &'static str)>

Get the bus index for this device in the current program.

pub fn is_strip(&self) -> bool

Check if this device is a strip

pub fn is_bus(&self) -> bool

Check if this device is a bus